perf(gemma-attn): fp32 split partials; MTP acceptance protocol forensics
- fused decode writes fp32 tmp_out partials (bf16 partial rounding was a
  candidate for greedy-stream forks; kept as strictly-better numerics);
  combine v1/v2 templated on partial type (legacy bf16 revert paths keep
  working); launcher checks tmp_out dtype float32; backend buffers float32
- units at/below prior floor (2.5e-4 -> 1.7e-4), gate unchanged,
  b32 E2E 1907.1 vs 1882.8 (+1.3%)
- forensics: the -12% MTP A/B was ~65% post-EOS junk (ignore_eos past the
  natural stop = OOD regime; both backends fork deterministically there).
  Fixed protocol (output >> max_tokens): real gap = -3.9% tok/s,
  acceptance 0.812 vs 0.944; hd512 verify numerics exonerated by eager
  path-swap A/B. Removed the measurement-only env branch.

+++ vllm/v1/attention/backends/gemma_attn.py
@@ -102,8 +102,10 @@ def _get_decode_partition_buffers(
         max_logits = torch.zeros(
             ns, num_heads, mp, dtype=torch.float32, device=device
         )
+        # FP32 split partials: bf16 partials turned split-count transitions
+        # into batch-wide greedy-stream forks (killed MTP draft acceptance).
         tmp_out = torch.zeros(
-            ns, num_heads, mp, head_size, dtype=dtype, device=device
+            ns, num_heads, mp, head_size, dtype=torch.float32, device=device
         )
         cached = (exp_sums, max_logits, tmp_out)
         _DECODE_PARTITION_CACHE[key] = cached
